refactor: split app into 3 common, master, agent modules
