+++ volta-agent/src/main/java/com/volta/agent/http/HttpSender.java
@@ -1,4 +1,4 @@
-package com.volta.http;
+package com.volta.agent.http;
 
 import java.io.IOException;
 import java.net.URI;
