+++ volta-agent/src/test/java/com/volta/agent/engine/LoadEngineTest.java
@@ -1,11 +1,11 @@
-package com.volta.engine;
+package com.volta.agent.engine;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.junit.jupiter.api.Assertions.*;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
-import com.volta.http.HttpSender;
+import com.volta.agent.http.HttpSender;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
