+++ volta-common/src/main/java/com/volta/stats/StatsSnapshot.java
